requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.2
discord==1.7.3
discord-components==2.1.2
discord.py==1.7.3
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pyxivapi==0.5.1
requests==2.27.1
selenium==4.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
wsproto==1.1.0
yarl==1.7.2
